consolidate browser utils
