+++ src/utils/browser.ts
@@ -1,0 +1,7 @@
+export function isSafari() {
+  return /^((?!chrome|android).)*safari/i.test(navigator.userAgent);
+}
+
+export function isFirefox() {
+  return navigator.userAgent.toLowerCase().includes('firefox');
+}
