--- src/utils/isFirefox.ts
@@ -1,3 +1,0 @@
-export default function isFirefox() {
-  return navigator.userAgent.toLowerCase().includes('firefox');
-}
